refactor(core): migrar Double a BigDecimal para precisión volumétrica y financiera
Cambios realizados:
- Refactorización de Value Objects y Entidades de dominio para asegurar precisión decimal.
- Implementación de comparaciones mediante .compareTo() y operaciones aritméticas inmutables.
- Actualización de lógica de validación de capacidad en Locations para prevenir errores de redondeo en el cálculo de peso y volumen.
- Ajuste de comandos y excepciones de dominio para mantener consistencia en la capa de aplicación.

Motivación: Evitar la pérdida de precisión inherente al punto flotante binario en operaciones de misión crítica, cumpliendo con los estándares de auditoría de inventarios industriales.

+++ src/main/java/com/juanbenevento/wms/audit/domain/AuditLog.java
@@ -2,6 +2,7 @@ package com.juanbenevento.wms.audit.domain;
 
 import com.juanbenevento.wms.inventory.domain.model.StockMovementType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record AuditLog(
@@ -10,9 +11,9 @@ public record AuditLog(
         StockMovementType type,
         String sku,
         String lpn,
-        Double quantity,
-        Double oldQuantity,
-        Double newQuantity,
+        BigDecimal quantity,
+        BigDecimal oldQuantity,
+        BigDecimal newQuantity,
         String user,
         String reason
 ) {}
